+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.2.0 (unreleased)
 
+- Update hit time calculations to account for the `queueTime` field (#10)
 - Add more descriptive error tracking (#9)
 - Update error tracking event category and action (#9)